pr conversations solving, rename success to webhook, change author in composer
