+++ composer.json
@@ -1,29 +1,28 @@
 {
-  "name": "payout/payment",
-  "description": "Payout Payment Extension",
-  "require": {
-    "php": "~5.5.0|~5.6.0|~7.0.0|~7.1.0|~7.2.0|~7.3.0|~7.4.0"
-  },
-  "type": "magento2-module",
-  "version": "1.0.0",
-  "license": [
-    "OSL-3.0",
-    "AFL-3.0"
-  ],
-  "authors": [
-    {
-      "name": "Michal Polák",
-      "email": "michal@payout.one",
-      "homepage": "https://payout.one/sk/",
-      "role": "Deputy CEO"
-    }
-  ],
-  "autoload": {
-    "files": [
-      "registration.php"
+    "name": "payout/payment",
+    "description": "Payout Payment Extension",
+    "require": {
+        "php": "~5.5.0|~5.6.0|~7.0.0|~7.1.0|~7.2.0|~7.3.0|~7.4.0"
+    },
+    "type": "magento2-module",
+    "version": "2.0.0",
+    "license": [
+        "OSL-3.0",
+        "AFL-3.0"
     ],
-    "psr-4": {
-      "Payout\\Payment\\": ""
+    "authors": [
+        {
+            "name": "Payout a.s.",
+            "email": "tech@payout.one",
+            "homepage": "https://payout.one/"
+        }
+    ],
+    "autoload": {
+        "files": [
+            "registration.php"
+        ],
+        "psr-4": {
+            "Payout\\Payment\\": ""
+        }
     }
-  }
 }
